[SORT] add abstraction folder
