+++ server/chat-service/internal/application/abstraction/message_service.go
@@ -1,0 +1,8 @@
+package abstraction
+
+import "github.com/widcraft/chat-service/internal/application/dto"
+
+type MessageService interface {
+	SaveMessage(message *dto.MessageDto) error
+	GetMessages(roomIdx uint) ([]dto.MessageDto, error)
+}
