+++ server/chat-service/internal/application/abstraction/messenger_service.go
@@ -1,0 +1,12 @@
+package abstraction
+
+import (
+	"github.com/widcraft/chat-service/internal/application/dto"
+	"github.com/widcraft/chat-service/internal/port/driving"
+)
+
+type MessengerService interface {
+	Participate(client driving.MessengerClient)
+	Quit(client driving.MessengerClient)
+	SendMessage(message *dto.MessageDto) error
+}
